uiScreens: move bagl related functions to uiScreens_bagl file
